feat: BREAKING CHANGE: complete restructure of component for accessibility audit
