+++ draft-packages/menu/KaizenDraft/Menu/components/MenuContent.tsx
@@ -1,4 +1,5 @@
 import * as React from "react"
+import { withDeprecatedComponent } from "@kaizen/react-deprecate-warning"
 import styles from "./MenuContent.module.scss"
 
 const MenuContent = (props: { children: React.ReactNode }) => (
