+++ draft-packages/menu/KaizenDraft/Menu/components/MenuHeader.tsx
@@ -1,6 +1,7 @@
 import * as React from "react"
+import { withDeprecatedComponent } from "@kaizen/react-deprecate-warning"
 
-import styles from "./MenuContent.module.scss"
+import styles from "./MenuSection.module.scss"
 
 const MenuHeader = (props: { title: string }) => (
   <div className={styles.header}>
