+++ draft-packages/menu/KaizenDraft/Menu/components/MenuSection.module.scss
@@ -1,0 +1,150 @@
+@import "~@kaizen/design-tokens/sass/border";
+@import "~@kaizen/design-tokens/sass/typography";
+@import "~@kaizen/design-tokens/sass/color";
+@import "~@kaizen/design-tokens/sass/spacing";
+@import "~@kaizen/deprecated-component-library-helpers/styles/layout";
+
+$side-padding: $spacing-sm;
+$gutter: calc(#{$spacing-md} * 1 / 6);
+$padding: $spacing-xs;
+
+.menuSectionWrapper {
+  &:not(:last-of-type) {
+    &::after {
+      width: 100%;
+      height: 1px;
+      background-color: $border-solid-border-color;
+      content: "";
+      display: block;
+      margin: 5px 0;
+    }
+  }
+}
+
+.menuSection {
+  list-style: none;
+  padding-left: 0;
+  margin: 0;
+
+  .menuItemList {
+    list-style: none;
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+
+    & > li {
+      display: inherit;
+    }
+  }
+}
+
+.header {
+  padding: 10px $side-padding 5px;
+  color: $color-purple-800;
+  margin: 0 $gutter;
+  font-family: $typography-heading-6-font-family;
+  font-weight: $typography-heading-6-font-weight;
+  font-size: $typography-heading-6-font-size;
+  line-height: $typography-heading-6-line-height;
+  letter-spacing: $typography-heading-6-letter-spacing;
+  display: block;
+}
+
+.menuItem {
+  display: flex;
+  align-items: center;
+  flex: auto;
+  box-sizing: border-box;
+  background: none;
+  border: 2px transparent solid;
+  text-align: left;
+  padding: 6px calc(#{$side-padding} - 2px);
+  margin: 0 $padding;
+  min-height: calc(1.75 * #{$spacing-md});
+  border-radius: 4px;
+  font-family: $typography-paragraph-body-font-family;
+  font-weight: $typography-paragraph-body-font-weight;
+  font-size: $typography-paragraph-body-font-size;
+  line-height: $typography-paragraph-body-line-height;
+  letter-spacing: $typography-paragraph-body-letter-spacing;
+  text-decoration: none;
+  color: $color-purple-800;
+
+  cursor: default;
+
+  &:not(.menuItem--disabled) {
+    cursor: pointer;
+  }
+
+  &:hover {
+    text-decoration: none;
+  }
+
+  &:not(.menuItem--disabled):hover,
+  &:focus {
+    background: $color-blue-100;
+    color: $color-blue-500;
+
+    .menuItem__Icon {
+      color: $color-blue-500;
+    }
+
+    &.menuItem--destructive {
+      background: $color-red-100;
+      color: $color-red-600;
+
+      .menuItem__Icon {
+        color: $color-red-600;
+      }
+    }
+  }
+
+  &:focus-visible,
+  &.focus-visible {
+    border-color: $color-blue-500;
+    outline: none;
+  }
+
+  [dir="rtl"] & {
+    text-align: right;
+  }
+}
+
+.menuItem--disabled {
+  color: rgba($color-purple-800-rgb, 0.3);
+  .menuItem__Icon {
+    color: rgba($color-purple-800-rgb, 0.3);
+  }
+}
+
+.menuItem--active {
+  color: $color-blue-500;
+  font-weight: $typography-paragraph-bold-font-weight;
+}
+
+.menuItem--destructive {
+  color: $color-red-600;
+  .menuItem__Icon {
+    color: $color-red-600;
+  }
+}
+
+.menuItem--destructive.menuItem--disabled {
+  color: rgba($color-red-600-rgb, 0.3);
+  .menuItem__Icon {
+    color: rgba($color-red-600-rgb, 0.3);
+  }
+}
+
+.menuItem__Label {
+  flex-grow: 1;
+}
+
+.menuItem__Icon {
+  @include ca-margin($end: calc(#{$spacing-md} / 3));
+
+  display: flex;
+  align-self: flex-start;
+  transform: translateY(0.2em);
+  color: rgba($color-purple-800-rgb, 0.75);
+}
