+++ draft-packages/menu/KaizenDraft/Menu/components/MenuSeparator.tsx
@@ -1,5 +1,6 @@
 import { Divider } from "@kaizen/draft-divider"
 import * as React from "react"
+import { withDeprecatedComponent } from "@kaizen/react-deprecate-warning"
 
 const MenuSeparator = () => <Divider variant="menuSeparator" />
 MenuSeparator.displayName = "MenuSeparator"
