+++ draft-packages/menu/index.ts
@@ -1,8 +1,10 @@
 export { default as Menu } from "./KaizenDraft/Menu/Menu"
 export { default as StatelessMenu } from "./KaizenDraft/Menu/StatelessMenu"
-export { default as MenuContent } from "./KaizenDraft/Menu/components/MenuContent"
-export { default as MenuHeader } from "./KaizenDraft/Menu/components/MenuHeader"
+export { default as MenuSection } from "./KaizenDraft/Menu/components/MenuSection"
 export { default as MenuItem } from "./KaizenDraft/Menu/components/MenuItem"
-export { default as MenuSeparator } from "./KaizenDraft/Menu/components/MenuSeparator"
 export { MenuItemProps } from "./KaizenDraft/Menu/components/MenuItem"
 export { MenuProps } from "./KaizenDraft/Menu/Menu"
+
+export { default as MenuContent } from "./KaizenDraft/Menu/components/MenuContent"
+export { default as MenuHeader } from "./KaizenDraft/Menu/components/MenuHeader"
+export { default as MenuSeparator } from "./KaizenDraft/Menu/components/MenuSeparator"
