+++ draft-packages/title-block-zen/package.json
@@ -36,7 +36,7 @@
     "@kaizen/draft-avatar": "^1.8.6",
     "@kaizen/draft-badge": "^1.7.0",
     "@kaizen/draft-button": "^5.2.4",
-    "@kaizen/draft-menu": "^3.4.21",
+    "@kaizen/draft-menu": "3.4.21",
     "@kaizen/draft-select": "^1.21.19",
     "@kaizen/draft-tag": "^1.12.6",
     "@types/classnames": "^2.2.10",
